+++ client/src/components/Nav/SettingsTabs/index.ts
@@ -1,1 +1,2 @@
 export { default as General } from './General';
+export { ClearChatsButton } from './General';
