+++ client/src/components/Nav/index.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { useState, useEffect, useRef, useContext } from 'react';
+import { useState, useEffect, useRef, useContext, useCallback } from 'react';
 import NewChat from './NewChat';
 import Panel from '../svg/Panel';
 import Spinner from '../svg/Spinner';
@@ -100,12 +100,12 @@ export default function Nav({ navVisible, setNavVisible }) {
     }
   };
 
-  const moveToTop = () => {
+  const moveToTop = useCallback(() => {
     const container = containerRef.current;
     if (container) {
       scrollPositionRef.current = container.scrollTop;
     }
-  };
+  }, [containerRef, scrollPositionRef]);
 
   const nextPage = async () => {
     moveToTop();
